Use Dataupdatecoordinator for glances
