+++ homeassistant/components/glances/translations/en.json
@@ -4,8 +4,7 @@
             "already_configured": "Device is already configured"
         },
         "error": {
-            "cannot_connect": "Failed to connect",
-            "wrong_version": "Version not supported (2 or 3 only)"
+            "cannot_connect": "Failed to connect"
         },
         "step": {
             "user": {
@@ -20,16 +19,6 @@
                     "version": "Glances API Version (2 or 3)"
                 },
                 "title": "Setup Glances"
-            }
-        }
-    },
-    "options": {
-        "step": {
-            "init": {
-                "data": {
-                    "scan_interval": "Update frequency"
-                },
-                "description": "Configure options for Glances"
             }
         }
     }
